refactor mcp jvm dto parser
